+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixes
 - Block taking jinja2.runtime.Undefined into DatabricksAdapter ([#98](https://github.com/databricks/dbt-databricks/pull/98))
+- Avoid using Cursor.schema API when database is None ([#100](https://github.com/databricks/dbt-databricks/pull/100))
 
 ## dbt-databricks 1.0.3 (April 26, 2022)